requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-dotenv==1.0.1
langchain==0.2.5
langchain-huggingface==0.0.3

